+++ Meta/gn/secondary/Userland/Libraries/LibGfx/BUILD.gn
@@ -98,6 +98,7 @@ shared_library("LibGfx") {
     "ImageFormats/WebPLoaderLossless.cpp",
     "ImageFormats/WebPLoaderLossy.cpp",
     "ImageFormats/WebPWriter.cpp",
+    "ImageFormats/WebPWriterLossless.cpp",
     "ImmutableBitmap.cpp",
     "Painter.cpp",
     "Palette.cpp",
